+++ week1/changho/movie-changho/src/MoviePoster.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+import PropTypes from 'prop-types'
+
+
+class MoviePoster extends Component {
+	
+	static propTypes = {
+		title: PropTypes.string,
+		image: PropTypes.string 
+	}
+
+	render() {
+		
+		return (
+			<div>
+				<img src={this.props.image} alt={this.props.title} />
+			</div>
+		);
+	}
+}
+export default MoviePoster;
